feat: add initialization for modules
