+++ void-base/src/main/kotlin/io/void/util/ModuleInit.kt
@@ -1,0 +1,13 @@
+package io.void.util
+
+abstract class ModuleInit {
+    companion object {
+        val initializers = mutableSetOf<ModuleInit>()
+    }
+
+    init {
+        initializers.add(this)
+    }
+
+    abstract fun init()
+}
